refactoring(fraud): use Header interface
To make fraud package more generic `*ExtendedHeader` usage is replaced
by Header interface (from libs/header).
