+++ fraud/proof.go
@@ -8,7 +8,7 @@ import (
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/metric/global"
 
-	"github.com/celestiaorg/celestia-node/header"
+	"github.com/celestiaorg/celestia-node/libs/header"
 )
 
 var (
@@ -56,7 +56,7 @@ type Proof interface {
 	// Validate throws an error if some conditions don't pass and thus fraud proof is not valid.
 	// NOTE: header.ExtendedHeader should pass basic validation otherwise it will panic if it's
 	// malformed.
-	Validate(*header.ExtendedHeader) error
+	Validate(header.Header) error
 
 	encoding.BinaryMarshaler
 	encoding.BinaryUnmarshaler
