+++ fraud/testing.go
@@ -34,7 +34,7 @@ func (d *DummyService) Get(context.Context, ProofType) ([]Proof, error) {
 }
 
 type mockStore struct {
-	headers    map[int64]*header.ExtendedHeader
+	headers    map[int64]header.Header
 	headHeight int64
 }
 
@@ -42,7 +42,7 @@ type mockStore struct {
 // headers.
 func createStore(t *testing.T, numHeaders int) *mockStore {
 	store := &mockStore{
-		headers:    make(map[int64]*header.ExtendedHeader),
+		headers:    make(map[int64]header.Header),
 		headHeight: 0,
 	}
 
@@ -59,7 +59,7 @@ func createStore(t *testing.T, numHeaders int) *mockStore {
 	return store
 }
 
-func (m *mockStore) GetByHeight(_ context.Context, height uint64) (*header.ExtendedHeader, error) {
+func (m *mockStore) GetByHeight(_ context.Context, height uint64) (header.Header, error) {
 	return m.headers[int64(height)], nil
 }
 
@@ -101,7 +101,7 @@ func (m *mockProof) Height() uint64 {
 	return 1
 }
 
-func (m *mockProof) Validate(*header.ExtendedHeader) error {
+func (m *mockProof) Validate(header.Header) error {
 	if !m.Valid {
 		return errors.New("mockProof: proof is not valid")
 	}
